requirements.txt
attrs==23.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
